edit: util에서도 getToday 제거
