+++ src/common/utils.ts
@@ -35,15 +35,6 @@ export const getDate = (date: any) => {
 };
 
 // 오늘 날짜 구하기 - 분양 정보 DB의 날짜 형식 0000-00-00
-export const getToday = () => {
-  const date = new Date();
-  const year = date.getFullYear();
-  const month = ('0' + (date.getMonth() + 1)).slice(-2);
-  const day = ('0' + date.getDate()).slice(-2);
-  const today = year + '-' + month + '-' + day;
-
-  return today;
-};
 
 // Alert(react confirm Library)실행함수
 export const customAlert = (text: string) => {
